First round of docstring for the Longformer utils
