--- speechbrain/nnet/attention_utilities/README.md
@@ -1,18 +1,0 @@
-# Utility functions for the Longformer attention mechanism
-All the code comes from: https://github.com/allenai/longformer and some change were done to fit SpeechBrain's architecture.
-
-    Longformer is an open-source project developed by the Allen Institute for Artificial Intelligence (AI2).
-    AI2 is a non-profit institute with the mission to contribute to humanity through high-impact AI research and
-    engineering.
-
-    The Longformer paper:
-        @article{
-        Beltagy2020Longformer,
-        title={Longformer: The Long-Document Transformer},
-        author={Iz Beltagy and Matthew E. Peters and Arman Cohan},
-        journal={arXiv:2004.05150},
-        year={2020}
-        }
-
-    Parts of the code found herein were modified by: Jonathan Tremblay (jonathan.tremblay.11@gmail.com) in order
-    to fit SpeechBrain's interface.
