+++ speechbrain/nnet/attention_utils/README.md
@@ -1,0 +1,2 @@
+# Utility functions for attention mechanisms
+The modules herein are utilities for different attention based models. Refer to the modules for more information.
